ai-services/resume-analyzer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-docx==1.1.0
PyPDF2==3.0.1
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
